# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/exercises/{id}:
    get:
      tags:
      - exercise-rest-controller
      operationId: getExerciseById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExerciseDTO"
    put:
      tags:
      - exercise-rest-controller
      operationId: updateExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExercisePutDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - exercise-rest-controller
      operationId: deleteExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/exercises/{id}/pdf:
    get:
      tags:
      - exercise-rest-controller
      operationId: getExercisePdf
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - exercise-rest-controller
      operationId: uploadExercisePdf
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pdfFile:
                  type: string
                  format: binary
              required:
              - pdfFile
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - exercise-rest-controller
      operationId: deleteExercisePdf
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/exercises/{id}/solutions/:
    post:
      tags:
      - exercise-rest-controller
      operationId: createSolution
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionPostDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/exerciselists/{id}/exercises/:
    post:
      tags:
      - exercise-list-rest-controller
      operationId: postExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExercisePostDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /exercise/{exerciseId}/pdf:
    get:
      tags:
      - exercise-web-controller
      operationId: downloadExercisePdf
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
...
  /api/v1/exercises/:
    get:
      tags:
      - exercise-rest-controller
      operationId: getExercises
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: listId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: nameFilter
        in: query
        required: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelExerciseDTO"
...
components:
  schemas:
...
    ExercisePutDTO:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
...
    SolutionPostDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ExercisePostDTO:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
...
    ExerciseListBasicInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
    UserBasicInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    SolutionBasicInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    ExerciseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        numSolutions:
          type: integer
          format: int32
        owner:
          $ref: "#/components/schemas/UserBasicInfoDTO"
        exerciseList:
          $ref: "#/components/schemas/ExerciseListBasicInfoDTO"
        solutions:
          type: array
          items:
            $ref: "#/components/schemas/SolutionBasicInfoDTO"
    PagedModelExerciseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...